Updating files to new structure
